Improve HTTP request: - calculate contentLength and contentType - add Basic Authentication - return failure when response is not successful
