+++ myExpenses/src/main/java/org/totschnig/myexpenses/util/io/NetworkUtils.kt
@@ -1,15 +1,24 @@
 package org.totschnig.myexpenses.util.io
 
 import android.app.Service
+import android.content.ContentResolver
 import android.content.Context
 import android.net.ConnectivityManager
 import android.net.NetworkCapabilities
+import android.net.Uri
 import android.net.wifi.WifiManager
 import android.os.Build
+import android.provider.OpenableColumns
 import android.text.format.Formatter
+import android.webkit.MimeTypeMap
 import androidx.annotation.IntRange
 import androidx.annotation.RequiresApi
+import dagger.internal.Preconditions
+import timber.log.Timber
+import java.io.File
 import java.net.Inet4Address
+
+const val MIME_TYPE_OCTET_STREAM = "application/octet-stream"
 
 fun isConnectedWifi(context: Context) = getConnectionType(context) == 2
 
@@ -62,3 +71,43 @@ private fun getWifiIpAddressLegacy(context: Context) =
             it
         )
     }
+
+fun calculateSize(contentResolver: ContentResolver, uri: Uri): Long {
+    val size: Long
+    Timber.d("Uri %s", uri)
+    if ("file" == uri.scheme) {
+        size = uri.path?.let { File(it) }?.length() ?: 0
+    } else {
+        contentResolver.query(uri, null, null, null, null).use { c ->
+            size = if (c != null) {
+                c.moveToFirst()
+                c.getLong(c.getColumnIndexOrThrow(OpenableColumns.SIZE))
+            } else {
+                -1
+            }
+        }
+    }
+    Timber.d("Size %d", size)
+    return size
+}
+
+fun getMimeType(fileName: String): String {
+    val result = MimeTypeMap.getSingleton().getMimeTypeFromExtension(getFileExtension(fileName))
+    return result ?: MIME_TYPE_OCTET_STREAM
+}
+
+//from Guava
+fun getNameWithoutExtension(file: String): String {
+    Preconditions.checkNotNull(file)
+    val fileName = File(file).name
+    val dotIndex = fileName.lastIndexOf('.')
+    return if (dotIndex == -1) fileName else fileName.substring(0, dotIndex)
+}
+
+//from Guava
+fun getFileExtension(fullName: String): String {
+    Preconditions.checkNotNull(fullName)
+    val fileName = File(fullName).name
+    val dotIndex = fileName.lastIndexOf('.')
+    return if (dotIndex == -1) "" else fileName.substring(dotIndex + 1)
+}
